Introduce a reverse_dns stage to lookup hostnames
Where a log record is missing a hostname (due to limitations at the
point of collection), being able to do a reverse DNS lookup for the IP
address to find the relevant hostnames can be a relevant substitute.

This version is restricted to using the systems default resolver,
further work could allow customization to (optionally) use a specified
resolver instead.